de objeto viaje se agrego : model, reposirity, service, controller, y vistas funcionando pero sin interfas de nav

+++ src/main/java/uber/clon/uber/Repository/ViajeRepository.java
@@ -5,6 +5,6 @@ import uber.clon.uber.Models.Viaje;
 
 import java.util.List;
 
-public interface ViajeRepository extends JpaRepository<Viaje, Long> {
+public interface ViajeRepository extends JpaRepository<Viaje, Integer> {
     List<Viaje> findByEstadoTrue(); // ✅ Solo devuelve viajes activos
 }
